Update package by adding seralization and app fee
git-tfs-id: [http://tfs2010:8080/tfs/gd]$/Delta/Services/AtlantisFramework/4.0;C371924

+++ Providers/Domains/DomainProductPackage/Atlantis.Framework.Providers.DomainProductPackage/PackageItems/ProductPackageItem.cs
@@ -10,7 +10,7 @@ namespace Atlantis.Framework.Providers.DomainProductPackage.PackageItems
   {
     private readonly Lazy<ICurrencyProvider> _currencyProvider;
 
-    public static ProductPackageItem Create(string name, int productId, int quantity, int duration, IProviderContainer container)
+    public static ProductPackageItem Create(string name, int productId, int quantity, double duration, IProviderContainer container)
     {
       var productPackageItem = new ProductPackageItem(name, container) {ProductId = productId, Duration = duration, Quantity = quantity};
